facts: allow resolving device links in lvm scenario
ceph-volume currently does not support /dev/disk paths but it is good practice not to rely on /dev/sd* paths and use /dev/disk/by-* format.
This allows the playbook to resolve these symlinks and correctly set device facts.
Also moves the tasks in a single block using the same when statement.